Restrict scanForTestClasses=false reproducer to JUnit 4
JUnit Platform (used by JUnit Vintage) returns null from getDetector(), so
DefaultTestScanner already takes the filenameScan path even with
scanForTestClasses=true. The reproducer's first assertion that the build
fails without the flag does not hold under Vintage. Move the test from
the abstract base into the JUnit 4-only concrete class.
